pyproject.toml
[tool.poetry]
name = "hazelnut"
version = "0.1.0"
description = ""
authors = ["Ethan Brooks <ethanabrooks@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
torch = [
    {platform = 'darwin', url =  "https://download.pytorch.org/whl/cpu/torch-1.11.0-cp38-none-macosx_10_9_x86_64.whl"},
    {platform = 'linux', url = "https://download.pytorch.org/whl/cu115/torch-1.11.0%2Bcu115-cp38-cp38-linux_x86_64.whl"}
]
numpy = "^1.22.3"
run-logger = "^0.1.0"
atari-py = "^0.3.0"
ale-py = "0.7.4"
stable-baselines3 = "^1.5.0"
gym = {version = "0.21.0", extras = ["atari"]}
AutoROM = {version = "^0.4.2", extras = ["accept-rom-license"]}
opencv-python = "^4.5.5"
torch-scatter = [
    {platform = 'darwin', url =  "https://data.pyg.org/whl/torch-1.11.0%2Bcpu/torch_scatter-2.0.9-cp38-cp38-macosx_10_14_x86_64.whl"},
    {platform = 'linux', url = "https://data.pyg.org/whl/torch-1.11.0%2Bcu115/torch_scatter-2.0.9-cp38-cp38-linux_x86_64.whl"}
]
torch-sparse = [
    {platform = 'darwin', url =  "https://data.pyg.org/whl/torch-1.11.0%2Bcpu/torch_sparse-0.6.13-cp38-cp38-macosx_10_14_x86_64.whl"},
    {platform = 'linux', url = "https://data.pyg.org/whl/torch-1.11.0%2Bcu115/torch_sparse-0.6.13-cp38-cp38-linux_x86_64.whl"}
]
torch-geometric = "^2.0.4"
GitPython = "^3.1.27"

[tool.poetry.dev-dependencies]
ipython = "^8.0.1"
ipdb = "^0.13.9"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.vscode
  | \__pycache__
  | \venv
  | _build
)/
'''

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
